pulled shooting out of main controller into ShootingController
