+++ README.md
@@ -1,5 +1,4 @@
 # Awesome Game Project v2
------------
 Lets nail it this time
 
 # Usage
@@ -8,6 +7,7 @@ Lets nail it this time
 npm install
 npm run build
 npm start
+visit --> http://localhost:3000/
 ```
 
 # Testing
